Fixed slice/entry issues; Enhanced dev setup
I was unable to use getters due the formatCurrencyString's reliance on state.currency and state.language. These values are revoked after the Immer is done with the draft state.

## `core/Entry.js`
* Created updateFormattedTotalPrice and updateFormattedValue helper functions
* Avoided passing currency and language down individually
* Avoided both spreading state and passing it down as a prop to each entry function
* Removed duplicate creation of the id value
* Removed unnecessary if statements for price_data and product_data; `{...undefined}` is valid code
* Converted the rest of the entry code to using Immer draft style
* Shortened `x = x + y` and similar to `x += y`
* Changed `updateQuantity` to use `updateEntry` under the hood like it used to
* Fixed quantity/count logic in `updateQuantity`
* Added radix to `parseInt`

## `core/slice.js`
* Formatting
* Remove `redirectToCheckout` and `checkoutSingleItem` from the reducers but still add them to the actions

## `core/store.js`
* A little formatting
* Generate the initial formatted price instead of guessing that they're using USD and English
* Using `isClient` now

## `utilities/serverless.js`
* Formatting
* Move spreading objects inline because `{ ...undefined }` and `{ ...{} }` are still valid

## `.eslintrc`
Added more snake case words we're using to ignore in the camelCase rule

I also fixed the CRA example's package.json so that it will now use our local use-shopping-cart instead of the published one. Make sure to run `yarn` to see this change.

## `examples/cra/functions/create-session.js`
* Added real error handling with status codes and messages

## `netlify.toml` and `package.json`
I edited these files to make the development process easier. I was having some issues with netlify dev so I made the configuration more specific. I also made it so that it won't open localhost:3000 (it really confused me) but will open localhost:8888.